# Performance optimizations for SvgColourConverter parser