… Source/Painting/SvgDeferredPaintServerFactory.cs
… Source/Svg.sln
… Tests/Svg.Benchmark/Program.cs
… Tests/Svg.Benchmark/README.md
… Tests/Svg.Benchmark/SpanBenchmarks.cs
… Tests/Svg.Benchmark/StringEqualityBenchmarks.cs
… Tests/Svg.Benchmark/SvgColourConverterBenchmarks.cs
… Tests/Svg.Benchmark/SvgPaintServerFactoryBenchmarks.cs
… Tests/Svg.Benchmark/ToLowerInvariantBenchmarks.cs
… Tests/Svg.UnitTests/LexerIssueTests.cs
… Tests/Svg.UnitTests/SvgColourConverterTests.cs
… Tests/Svg.UnitTests/SvgColourConverterTests_OLD.cs