@wieslawsoltes 2021-01-15
|                            Method |        Job |       Runtime |      Mean |    Error |   StdDev |         Op/s | Ratio | Rank |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|---------------------------------- |----------- |-------------- |----------:|---------:|---------:|-------------:|------:|-----:|-------:|------:|------:|----------:|
|            StringEqualityOperator | Job-WTSXZF |    .NET 4.6.1 |  79.29 ns | 0.352 ns | 0.294 ns | 12,611,356.2 |  1.00 |    1 | 0.0134 |     - |     - |      56 B |
|     MemoryExtensionsSequenceEqual | Job-WTSXZF |    .NET 4.6.1 | 130.39 ns | 0.645 ns | 0.603 ns |  7,669,315.5 |  1.65 |    2 | 0.0134 |     - |     - |      56 B |
| MemoryExtensionsSequenceCompareTo | Job-WTSXZF |    .NET 4.6.1 | 138.51 ns | 0.590 ns | 0.523 ns |  7,219,484.1 |  1.75 |    3 | 0.0134 |     - |     - |      56 B |
|         MemoryExtensionsCompareTo | Job-WTSXZF |    .NET 4.6.1 | 129.68 ns | 0.567 ns | 0.530 ns |  7,711,058.0 |  1.64 |    2 | 0.0134 |     - |     - |      56 B |
|                SpanManualEquality | Job-WTSXZF |    .NET 4.6.1 | 130.74 ns | 0.501 ns | 0.444 ns |  7,648,880.3 |  1.65 |    2 | 0.0134 |     - |     - |      56 B |
|                                   |            |               |           |          |          |              |       |      |        |       |       |           |
|            StringEqualityOperator | Job-VCPUFP | .NET Core 2.1 |  40.58 ns | 0.304 ns | 0.284 ns | 24,641,255.3 |  1.00 |    1 | 0.0133 |     - |     - |      56 B |
|     MemoryExtensionsSequenceEqual | Job-VCPUFP | .NET Core 2.1 |  47.40 ns | 0.215 ns | 0.201 ns | 21,095,445.9 |  1.17 |    2 |      - |     - |     - |         - |
| MemoryExtensionsSequenceCompareTo | Job-VCPUFP | .NET Core 2.1 |  48.50 ns | 0.111 ns | 0.092 ns | 20,618,269.0 |  1.19 |    3 |      - |     - |     - |         - |
|         MemoryExtensionsCompareTo | Job-VCPUFP | .NET Core 2.1 |  58.85 ns | 0.287 ns | 0.268 ns | 16,992,626.8 |  1.45 |    5 |      - |     - |     - |         - |
|                SpanManualEquality | Job-VCPUFP | .NET Core 2.1 |  49.23 ns | 0.136 ns | 0.127 ns | 20,312,715.5 |  1.21 |    4 |      - |     - |     - |         - |
|                                   |            |               |           |          |          |              |       |      |        |       |       |           |
|            StringEqualityOperator | Job-NZZZER | .NET Core 3.1 |  30.36 ns | 0.133 ns | 0.118 ns | 32,938,472.7 |  1.00 |    1 | 0.0114 |     - |     - |      48 B |
|     MemoryExtensionsSequenceEqual | Job-NZZZER | .NET Core 3.1 |  36.15 ns | 0.124 ns | 0.104 ns | 27,664,440.6 |  1.19 |    2 |      - |     - |     - |         - |
| MemoryExtensionsSequenceCompareTo | Job-NZZZER | .NET Core 3.1 |  38.71 ns | 0.097 ns | 0.086 ns | 25,831,061.6 |  1.28 |    3 |      - |     - |     - |         - |
|         MemoryExtensionsCompareTo | Job-NZZZER | .NET Core 3.1 |  43.76 ns | 0.210 ns | 0.186 ns | 22,850,792.4 |  1.44 |    5 |      - |     - |     - |         - |
|                SpanManualEquality | Job-NZZZER | .NET Core 3.1 |  41.29 ns | 0.078 ns | 0.065 ns | 24,217,657.4 |  1.36 |    4 |      - |     - |     - |         - |
|                                   |            |               |           |          |          |              |       |      |        |       |       |           |
|            StringEqualityOperator | Job-JRWFWN | .NET Core 5.0 |  27.96 ns | 0.110 ns | 0.103 ns | 35,763,406.5 |  1.00 |    4 | 0.0115 |     - |     - |      48 B |
|     MemoryExtensionsSequenceEqual | Job-JRWFWN | .NET Core 5.0 |  23.86 ns | 0.050 ns | 0.046 ns | 41,904,046.9 |  0.85 |    1 |      - |     - |     - |         - |
| MemoryExtensionsSequenceCompareTo | Job-JRWFWN | .NET Core 5.0 |  25.46 ns | 0.056 ns | 0.047 ns | 39,269,733.6 |  0.91 |    2 |      - |     - |     - |         - |
|         MemoryExtensionsCompareTo | Job-JRWFWN | .NET Core 5.0 |  28.08 ns | 0.044 ns | 0.041 ns | 35,611,421.0 |  1.00 |    4 |      - |     - |     - |         - |
|                SpanManualEquality | Job-JRWFWN | .NET Core 5.0 |  27.12 ns | 0.110 ns | 0.098 ns | 36,873,957.9 |  0.97 |    3 |      - |     - |     - |         - |