@wieslawsoltes 2021-01-15
Same performance, no allocations for system colors.

`dotnet run -c Release -f netcoreapp3.1 -- -f '*SvgColourConverter_Parse_system_colors_single_DIRECT*'`
```
|                                                          Method |     Mean |    Error |   StdDev |         Op/s | Rank |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|---------------------------------------------------------------- |---------:|---------:|---------:|-------------:|-----:|-------:|------:|------:|----------:|
|      SvgColourConverter_Parse_system_colors_single_DIRECT_first | 58.21 ns | 0.229 ns | 0.203 ns | 17,178,056.4 |    2 |      - |     - |     - |         - |
|     SvgColourConverter_Parse_system_colors_single_DIRECT_middle | 60.31 ns | 0.331 ns | 0.294 ns | 16,581,537.3 |    3 |      - |     - |     - |         - |
|       SvgColourConverter_Parse_system_colors_single_DIRECT_last | 56.39 ns | 0.181 ns | 0.170 ns | 17,735,183.0 |    1 |      - |     - |     - |         - |
|  SvgColourConverter_Parse_system_colors_single_DIRECT_first_OLD | 60.27 ns | 0.404 ns | 0.337 ns | 16,592,354.7 |    3 | 0.0114 |     - |     - |      48 B |
| SvgColourConverter_Parse_system_colors_single_DIRECT_middle_OLD | 63.84 ns | 0.827 ns | 0.773 ns | 15,665,051.6 |    4 | 0.0134 |     - |     - |      56 B |
|   SvgColourConverter_Parse_system_colors_single_DIRECT_last_OLD | 59.94 ns | 0.212 ns | 0.188 ns | 16,683,379.1 |    3 | 0.0114 |     - |     - |      48 B |
```