@inforithmics 2021-01-16
Looking at your benchmark results on the Pull request

SvgColourConverter_Parse_system_colors_single_DIRECT_first seems to be slower in than old implementation
But you have probable already improved the performance since measuring it. So I wonder what are the actual Performance numbers now.