@wieslawsoltes 2021-01-16
> Looking at your benchmark results on the Pull request
> 
> SvgColourConverter_Parse_system_colors_single_DIRECT_first seems to be slower in than old implementation
> But you have probable already improved the performance since measuring it. So I wonder what are the actual Performance numbers now.

That is special case and really hard to make it faster than original `switch (string)` as compiler does a really great jog here, but it requires an additional allocation for make string lowercase. My solution does not allocate but it's a bit slower (about 25%). This is only case for SystemColors, that are not that much used (I think) and I have moved those checks as last so they are not so often executed.