@inforithmics 2021-01-19
> > Looking at your benchmark results on the Pull request
> > SvgColourConverter_Parse_system_colors_single_DIRECT_first seems to be slower in than old implementation
> > But you have probable already improved the performance since measuring it. So I wonder what are the actual Performance numbers now.
> 
> That is special case and really hard to make it faster than original `switch (string)` as compiler does a really great jog here, but it requires an additional allocation for make string lowercase. My solution does not allocate but it's a bit slower (about 25%). This is only case for SystemColors, that are not that much used (I think) and I have moved those checks as last so they are not so often executed.
> 
> I will do benchmarks later, I think net5.0 has some optimizations for that case, I have done a lot to make it fast. The ReadOnlySpan unfortunately does not have switch statement support yet.

Would it be possible to Make the Lookup Dictionary CaseInsensitive and avoid the ToLowerAscii call?

var caseInsensitiveDictionary = new Dictionary<string,...>(StringComparer.OrdinalIgnoreCase);